feat: Add weather units parameter and update data models to replace Double and Int for String. In order to simplify the database a `Daily` class has been added to represent a simplified daily forecast.
This commit includes the following changes:

-   **`WeatherUnits` Enum**: Adds a new `WeatherUnits` enum (`STANDARD`, `METRIC`, `IMPERIAL`) to the networking module.
-   **API Service Update**:
    -   Modifies `WeatherApiService` to accept `WeatherUnits` in its constructor.
    -   Updates the API request URL to include the `units` query parameter.
    -   Injects `WeatherUnits.STANDARD` as the default unit in the `NetworkingModule`.
-   **Data Model Refactoring**:
    -   Updates the data types in networking models (`DailyData`, `TempData`, `FeelsLikeData`, etc.) from numeric types like `Int`, `Double` to `String?` or `Long?`. This change makes the models more robust to handle varied formats from the API.
-   **Domain Model**: Introduces a new `Daily` data class in the `domain` module to represent a simplified daily forecast.

+++ feature-weather/networking/src/commonMain/kotlin/com/charly/networking/model/DailyForecastWeatherData.kt
@@ -13,7 +13,7 @@ data class DailyForecastWeatherData(
     @SerialName("timezone")
     val timezone: String? = null,
     @SerialName("timezone_offset")
-    val timezoneOffset: Int? = null,
+    val timezoneOffset: Long? = null,
     @SerialName("daily")
     val daily: List<DailyData> = emptyList()
 )
